+++ app/src/main/AndroidManifest.xml
@@ -2,7 +2,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
 
-    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
     <uses-permission android:name="android.permission.INTERNET"/>
 
     <application
@@ -16,7 +15,7 @@
         android:theme="@style/Theme.PlaylistMaker"
         tools:targetApi="31">
         <activity
-            android:name=".main.MainActivity"
+            android:name=".presentation.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -24,7 +23,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".settings.ui.activiy.SettingsActivity"
+            android:name=".presentation.SettingsActivity"
             android:exported="false">
         </activity>
         <activity
@@ -32,12 +31,11 @@
             android:exported="false">
         </activity>
         <activity
-            android:name=".search.ui.SearchActivity"
+            android:name=".presentation.SearchActivity"
             android:exported="false">
         </activity>
         <activity
-            android:screenOrientation="portrait"
-            android:name=".player.AudioPlayer"
+            android:name=".presentation.AudioPlayer"
             android:exported="false">
         </activity>
