+++ app/src/main/java/com/example/playlistmaker/creator/Creator.kt
@@ -1,17 +1,21 @@
 package com.example.playlistmaker.creator
 
 import android.content.Context
-import com.example.playlistmaker.data.Player
-import com.example.playlistmaker.data.SearchHistory
-import com.example.playlistmaker.data.TracksRepositoryImpl
-import com.example.playlistmaker.data.network.RetrofitNetworkClient
-import com.example.playlistmaker.domain.PlayerPresenter
-import com.example.playlistmaker.domain.api.OneTrackRepository
-import com.example.playlistmaker.domain.api.TrackHistoryRepository
-import com.example.playlistmaker.domain.api.TracksInteractor
-import com.example.playlistmaker.domain.api.TracksRepository
-import com.example.playlistmaker.domain.impl.TracksInteractorImpl
-import com.example.playlistmaker.domain.use_case.PlayControlImpl
+import com.example.playlistmaker.player.data.Player
+import com.example.playlistmaker.search.data.SearchHistory
+import com.example.playlistmaker.search.data.TracksRepositoryImpl
+import com.example.playlistmaker.search.data.network.RetrofitNetworkClient
+import com.example.playlistmaker.player.domain.PlayerPresenter
+import com.example.playlistmaker.search.domain.api.OneTrackRepository
+import com.example.playlistmaker.search.domain.api.TrackHistoryRepository
+import com.example.playlistmaker.search.domain.api.TracksInteractor
+import com.example.playlistmaker.search.domain.api.TracksRepository
+import com.example.playlistmaker.search.domain.impl.TracksInteractorImpl
+import com.example.playlistmaker.player.domain.use_case.PlayControlImpl
+import com.example.playlistmaker.sharing.data.ExternalNavigator
+import com.example.playlistmaker.sharing.data.impl.ExternalNavigatorImpl
+import com.example.playlistmaker.sharing.domain.SharingInteractor
+import com.example.playlistmaker.sharing.domain.impl.SharingInteractorImpl
 
 
 object Creator {
@@ -33,6 +37,14 @@ object Creator {
 
     fun getHistoryRepository(context: Context): TrackHistoryRepository {
         return SearchHistory(context)
+    }
+
+    fun getSharingInteractor(context: Context):SharingInteractor {
+        return SharingInteractorImpl(getExternalNavigator(),context)
+    }
+
+    fun getExternalNavigator(): ExternalNavigator {
+        return ExternalNavigatorImpl()
     }
 
 }
