--- app/src/main/java/com/example/playlistmaker/data/InputStatus.kt
@@ -1,6 +1,0 @@
-package com.example.playlistmaker
-
-enum class InputStatus {
-    SUCCESS, EMPTY, ERROR
-}
-
