--- app/src/main/java/com/example/playlistmaker/data/NetworkException.kt
@@ -1,4 +1,0 @@
-package com.example.playlistmaker.data
-
-class NetworkException() : Exception() {
-}
