--- app/src/main/java/com/example/playlistmaker/data/TracksRepositoryImpl.kt
@@ -1,34 +1,0 @@
-package com.example.playlistmaker.data
-
-import com.example.playlistmaker.data.dto.SearchRequest
-import com.example.playlistmaker.data.dto.TrackResponse
-import com.example.playlistmaker.domain.api.TracksRepository
-import com.example.playlistmaker.domain.models.Track
-
-class TracksRepositoryImpl(private val networkClient: NetworkClient) : TracksRepository {
-
-    override fun searchTracks(expression: String): List<Track> {
-        val response = try {
-            networkClient.doRequest(SearchRequest(expression))
-        } catch (e: Exception) {
-            throw NetworkException()
-        }
-        if (response.resultCode == 200) {
-            return (response as TrackResponse).results.map {
-                Track(
-                    it.trackName,
-                    it.artistName,
-                    it.trackTimeMillis,
-                    it.artworkUrl100,
-                    it.collectionName,
-                    it.releaseDate,
-                    it.primaryGenreName,
-                    it.country,
-                    it.previewUrl
-                )
-            }
-        } else {
-            return emptyList()
-        }
-    }
-}
