--- app/src/main/java/com/example/playlistmaker/data/dto/SearchRequest.kt
@@ -1,3 +1,0 @@
-package com.example.playlistmaker.data.dto
-
-data class SearchRequest(val expression: String)
