--- app/src/main/java/com/example/playlistmaker/data/dto/TrackResponse.kt
@@ -1,8 +1,0 @@
-package com.example.playlistmaker.data.dto
-
-
-data class TrackResponse(
-    val resultCount: Int,
-    val expression: String,
-    val results: List<TrackDto>
-) : Response()
