--- app/src/main/java/com/example/playlistmaker/data/network/ApiSearch.kt
@@ -1,12 +1,0 @@
-package com.example.playlistmaker.data.network
-
-import com.example.playlistmaker.data.dto.TrackResponse
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface ApiSearch {
-    @GET("/search?entity=song")
-    fun search(@Query("term") text: String): Call<TrackResponse>
-
-}
