--- app/src/main/java/com/example/playlistmaker/domain/PlayerPresenter.kt
@@ -1,10 +1,0 @@
-package com.example.playlistmaker.domain
-
-interface PlayerPresenter {
-    fun startPlayer()
-    fun pausePlayer()
-    fun progressTimeViewUpdate(progressTime: String)
-    fun playButtonEnabled()
-    fun postDelayed(runnable: Runnable)
-    fun removeCallbacks(runnable: Runnable)
-}
