--- app/src/main/java/com/example/playlistmaker/domain/api/OneTrackRepository.kt
@@ -1,7 +1,0 @@
-package com.example.playlistmaker.domain.api
-
-import com.example.playlistmaker.domain.models.Track
-
-interface OneTrackRepository {
-    fun getTrack(): Track
-}
