--- app/src/main/java/com/example/playlistmaker/domain/models/Track.kt
@@ -1,14 +1,0 @@
-package com.example.playlistmaker.domain.models
-
-
-data class Track(
-    val trackName: String,
-    val artistName: String,
-    val trackTimeMillis: Long,
-    val artworkUrl100: String,
-    val collectionName: String,
-    val releaseDate: String,
-    val primaryGenreName: String,
-    val country: String,
-    val previewUrl: String
-)
