+++ app/src/main/java/com/example/playlistmaker/player/domain/PlayerInteractor.kt
@@ -1,0 +1,12 @@
+package com.example.playlistmaker.player.domain
+
+import com.example.playlistmaker.search.domain.models.Track
+
+interface PlayerInteractor {
+
+    fun preparePlayer(track: Track, prepareCallback: () -> Unit, completeCallback: () -> Unit)
+    fun startPlayer()
+    fun pausePlayer()
+    fun getCurrentPosition(): Int
+    fun release()
+}
