+++ app/src/main/java/com/example/playlistmaker/player/domain/PlayerPresenter.kt
@@ -1,0 +1,10 @@
+package com.example.playlistmaker.player.domain
+
+interface PlayerPresenter {
+    fun startPlayer()
+    fun pausePlayer()
+    fun progressTimeViewUpdate(progressTime: String)
+    fun playButtonEnabled()
+    fun postDelayed(runnable: Runnable)
+    fun removeCallbacks(runnable: Runnable)
+}
