+++ app/src/main/java/com/example/playlistmaker/player/domain/use_case/PlayControl.kt
@@ -1,0 +1,11 @@
+package com.example.playlistmaker.player.domain.use_case
+
+import com.example.playlistmaker.search.domain.models.Track
+
+interface PlayControl {
+    fun playbackControl()
+    fun preparePlayer(item: Track)
+    fun pausePlayer()
+    fun createUpdateProgressTimeRunnable(): Runnable
+    fun release()
+}
