--- app/src/main/java/com/example/playlistmaker/presentation/LibraryActivity.kt
@@ -1,18 +1,0 @@
-package com.example.playlistmaker.presentation
-
-import android.os.Bundle
-import android.widget.ImageView
-import androidx.appcompat.app.AppCompatActivity
-import com.example.playlistmaker.R
-
-class LibraryActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_library)
-
-        val imageBack = findViewById<ImageView>(R.id.backToMainActivity)
-        imageBack.setOnClickListener { finish() }
-
-    }
-}
