+++ app/src/main/java/com/example/playlistmaker/search/data/InputStatus.kt
@@ -1,0 +1,6 @@
+package com.example.playlistmaker.search.data
+
+enum class InputStatus {
+    SUCCESS, EMPTY, ERROR
+}
+
