+++ app/src/main/java/com/example/playlistmaker/search/data/NetworkClient.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.search.data
+
+import com.example.playlistmaker.search.data.dto.Response
+
+interface NetworkClient {
+    fun doRequest(dto: Any): Response
+}
