+++ app/src/main/java/com/example/playlistmaker/search/data/NetworkException.kt
@@ -1,0 +1,4 @@
+package com.example.playlistmaker.search.data
+
+class NetworkException() : Exception() {
+}
