+++ app/src/main/java/com/example/playlistmaker/search/data/TracksRepositoryImpl.kt
@@ -1,0 +1,34 @@
+package com.example.playlistmaker.search.data
+
+import com.example.playlistmaker.search.data.dto.SearchRequest
+import com.example.playlistmaker.search.data.dto.TrackResponse
+import com.example.playlistmaker.search.domain.api.TracksRepository
+import com.example.playlistmaker.search.domain.models.Track
+
+class TracksRepositoryImpl(private val networkClient: NetworkClient) : TracksRepository {
+
+    override fun searchTracks(expression: String): List<Track> {
+        val response = try {
+            networkClient.doRequest(SearchRequest(expression))
+        } catch (e: Exception) {
+            throw NetworkException()
+        }
+        if (response.resultCode == 200) {
+            return (response as TrackResponse).results.map {
+                Track(
+                    it.trackName,
+                    it.artistName,
+                    it.trackTimeMillis,
+                    it.artworkUrl100,
+                    it.collectionName,
+                    it.releaseDate,
+                    it.primaryGenreName,
+                    it.country,
+                    it.previewUrl
+                )
+            }
+        } else {
+            return emptyList()
+        }
+    }
+}
