+++ app/src/main/java/com/example/playlistmaker/search/data/dto/TrackResponse.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker.search.data.dto
+
+
+data class TrackResponse(
+    val resultCount: Int,
+    val expression: String,
+    val results: List<TrackDto>
+) : Response()
