+++ app/src/main/java/com/example/playlistmaker/search/domain/api/OneTrackRepository.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.search.domain.api
+
+import com.example.playlistmaker.search.domain.models.Track
+
+interface OneTrackRepository {
+    fun getTrack(): Track
+}
