+++ app/src/main/java/com/example/playlistmaker/search/domain/api/TracksRepository.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.search.domain.api
+
+import com.example.playlistmaker.search.domain.models.Track
+
+interface TracksRepository {
+    fun searchTracks(expression: String): List<Track>
+}
