+++ app/src/main/java/com/example/playlistmaker/sharing/domain/model/EmailData.kt
@@ -1,4 +1,8 @@
 package com.example.playlistmaker.sharing.domain.model
 
-class EmailData {
-}
+
+data class EmailData(
+    val emailAddressee: Array<String>,
+    val emailTopic: String,
+    val emailMessage: String
+)
